% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wagepan_mgt}
\alias{wagepan_mgt}
\title{wagepan_mgt}
\format{
## `wagepan_mgt`
A data frame with 4,360 rows and 5 columns:
\describe{
  \item{lwage}{Log wage.}
  \item{nr}{Worker ID.}
  \item{year}{Year}
  \item{union}{Union status.}
  \item{hours}{Annual Hours worked.}
}
}
\source{
<http://fmwww.bc.edu/ec-p/data/wooldridge/wagepan.des>
}
\usage{
wagepan_mgt
}
\description{
A subset of data from Vella and Verbeek (1998).
}
\keyword{datasets}
